refactor(#458): separate create project dialog
